Created AppRunner class
Moved all logic from Main to AppRunner
